+++ .github/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+Update <small>_ February 2024</small>
+
+- fix: message edit and delete crash (01/02/2024)
+
 Update <small>_ January 2024</small>
 
 - fix: defer interaction for cache update (28/01/2024)